requirements_dev.txt
tox
pytest
pytest-cov
pytest-random-order
pytest-django
freezegun
flake8
black
